# doc/API接口/openapi.yaml
openapi: 3.0.3
info:
  title: 消息通知平台 API
  description: |
    消息通知平台提供统一的消息发送服务，支持多种渠道（邮件、短信、微信等）。
    
    ## 特性
    - 📧 邮件发送（支持HTML、附件、模板）
    - 📱 短信发送（第二期）
    - 💬 微信消息（第二期）
    - 📝 消息模板管理
    - 🔐 JWT Token认证
    - 📊 监控和健康检查
    
    ## 认证方式
    1. 使用API Key和API Secret获取JWT Token
    2. 后续请求携带Token：`Authorization: Bearer {token}`
    
    ## 限流策略
    - 每个API Key默认限制：1000次/小时
    - 超过限制返回429状态码
    
    ## 技术支持
    - 邮箱：support@example.com
    - 文档：https://docs.example.com
    
  version: 1.0.0
  contact:
    name: 技术支持团队
    email: support@example.com
    url: https://docs.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 本地开发环境
  - url: https://api-dev.example.com
    description: 开发环境
  - url: https://api-staging.example.com
    description: 测试环境
  - url: https://api.example.com
    description: 生产环境
...
paths:
...
  /api/v1/templates:
    get:
      tags:
        - 模板
      summary: 查询模板列表
      description: 分页查询模板列表
      operationId: listTemplates
      security:
        - bearerAuth: []
      parameters:
        - name: template_type
          in: query
          description: 模板类型
          schema:
            type: string
            enum: [email, sms, wechat]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListResponse'
    
    post:
      tags:
        - 模板
      summary: 创建模板
      description: 创建新的消息模板
      operationId: createTemplate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/templates/{template_id}:
    get:
      tags:
        - 模板
      summary: 查询模板详情
      description: 根据模板ID查询模板详细信息
      operationId: getTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - 模板
      summary: 更新模板
      description: 更新模板内容，会自动创建新版本
      operationId: updateTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - 模板
      summary: 删除模板
      description: 软删除模板（设置deleted_at）
      operationId: deleteTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/templates/{template_id}/history:
    get:
      tags:
        - 模板
      summary: 查询模板历史版本
      description: 查询模板的所有历史版本
      operationId: getTemplateHistory
      security:
        - bearerAuth: []
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateHistoryResponse'
  /api/v1/templates/{template_id}/rollback:
    post:
      tags:
        - 模板
      summary: 回滚模板
      description: 回滚到指定历史版本
      operationId: rollbackTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateRollbackRequest'
      responses:
        '200':
          description: 回滚成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
  /api/v1/templates/preview:
    post:
      tags:
        - 模板
      summary: 预览模板
      description: 使用测试数据渲染模板，查看最终效果
      operationId: previewTemplate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplatePreviewRequest'
      responses:
        '200':
          description: 预览成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatePreviewResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，0表示成功
          example: 0
        message:
          type: string
          description: 响应消息
          example: "success"
        request_id:
          type: string
          description: 请求追踪ID
          example: "req_xyz789"
        timestamp:
          type: string
          format: date-time
          description: 响应时间
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
    
    # ============ 健康检查 ============
...
    TemplateCreateRequest:
      type: object
      required:
        - template_name
        - template_type
        - content
      properties:
        template_name:
          type: string
          maxLength: 100
          example: "注册验证码"
        template_type:
          type: string
          enum: [email, sms, wechat]
          example: "email"
        subject:
          type: string
          maxLength: 255
          example: "您的验证码"
        content:
          type: string
          example: "<p>您的验证码是：{{code}}</p>"
        variables:
          type: array
          items:
            type: string
          example: ["code", "username"]
        description:
          type: string
          example: "用户注册验证码模板"
        created_by:
          type: string
          example: "admin"
    TemplateUpdateRequest:
      type: object
      properties:
        subject:
          type: string
        content:
          type: string
        variables:
          type: array
          items:
            type: string
        change_description:
          type: string
          example: "修改邮件主题"
        updated_by:
          type: string
    TemplateResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                template_name:
                  type: string
                template_type:
                  type: string
                subject:
                  type: string
                content:
                  type: string
                variables:
                  type: array
                  items:
                    type: string
                version:
                  type: integer
                is_active:
                  type: boolean
                description:
                  type: string
                created_at:
                  type: string
                  format: date-time
    TemplateDetailResponse:
      allOf:
        - $ref: '#/components/schemas/TemplateResponse'
    TemplateListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total:
                  type: integer
                page:
                  type: integer
                page_size:
                  type: integer
                records:
                  type: array
                  items:
                    type: object
    TemplateHistoryResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  version:
                    type: integer
                  content:
                    type: string
                  created_by:
                    type: string
                  created_at:
                    type: string
                    format: date-time
    TemplateRollbackRequest:
      type: object
      required:
        - version
      properties:
        version:
          type: integer
          description: 要回滚到的版本号
          example: 1
        change_description:
          type: string
          example: "回滚到初始版本"
        updated_by:
          type: string
          example: "admin"
    TemplatePreviewRequest:
      type: object
      required:
        - template_id
        - template_vars
      properties:
        template_id:
          type: integer
          example: 1
        template_vars:
          type: object
          additionalProperties: true
          example:
            username: "张三"
            code: "123456"
    TemplatePreviewResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                subject:
                  type: string
                  example: "您的验证码"
                content:
                  type: string
                  example: "<p>您的验证码是：123456</p>"
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1001
            message: "请求参数错误"
            data: null
            request_id: "req_xyz789"
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1004
            message: "资源不存在"
            data: null
            request_id: "req_xyz789"
security:
  - bearerAuth: []